refactor(research): extract visual text quality tests
